requirements.txt
anyio==3.6.1
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cookiecutter==2.1.1
deta==1.1.0
Django==4.0.5
django-auth-ldap==4.1.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
fastapi==0.78.0
h11==0.13.0
HTMLParser==0.0.2
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
Markdown==3.3.7
markup==0.2
MarkupSafe==2.1.1
numpy==1.23.0
orjson==3.7.7
pandas==1.4.3
Pillow==9.2.0
pipreqs==0.4.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
PyJWT==1.7.1
PySimpleGUI==4.60.4
PySimpleGUIWeb==0.39.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-ldap==3.4.0
python-multipart==0.0.5
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
remi==2020.3.10
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.19.1
text-unidecode==1.3
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
yarg==0.1.9
